app/requirements.txt
SQLAlchemy~=2.0.41
python-dotenv~=1.0.1
python-multipart~=0.0.20
pydantic-settings~=2.9.1
fastapi~=0.115.12
alembic~=1.15.2
pydantic~=2.11.4
pytest~=8.3.5
httpx~=0.28.1
psycopg2~=2.9.10
asyncpg~=0.30.0
uvicorn~=0.34.2